+++ simplesource-command-serialization/src/test/java/io/simplesource/kafka/serialization/avro/mappers/AvroCustomEventSerdeTests.java
@@ -4,26 +4,29 @@ import io.simplesource.api.CommandError;
 import io.simplesource.api.CommandId;
 import io.simplesource.data.Result;
 import io.simplesource.data.Sequence;
+import io.simplesource.kafka.api.AggregateSerdes;
+import io.simplesource.kafka.model.AggregateUpdate;
 import io.simplesource.kafka.model.CommandRequest;
 import io.simplesource.kafka.model.CommandResponse;
-import io.simplesource.kafka.serialization.avro.AvroCommandSerdes;
-import io.simplesource.kafka.serialization.avro.mappers.domain.UserAccountDomainCommand;
-import io.simplesource.kafka.serialization.avro.mappers.domain.UserAccountDomainKey;
+import io.simplesource.kafka.model.ValueWithSequence;
+import io.simplesource.kafka.serialization.avro.AvroSerdes;
+import io.simplesource.kafka.serialization.avro.mappers.domain.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-class AvroCommandSerdeTests {
+class AvroCustomEventSerdeTests {
     private static final String topic = "topic";
-    private AvroCommandSerdes<UserAccountDomainKey, UserAccountDomainCommand> serdes;
+    private AggregateSerdes<UserAccountDomainKey, UserAccountDomainEvent, UserAccountDomainEvent, Boolean> serdes;
 
     @BeforeEach
     void setup() {
-        serdes = AvroCommandSerdes.of(
+        serdes = AvroSerdes.Custom.event(
                 UserAccountAvroMappers.keyMapper,
-                UserAccountAvroMappers.commandMapper,
+                UserAccountAvroMappers.eventMapper,
                 "http://localhost:8081",
                 true);
     }
@@ -42,28 +45,48 @@ class AvroCommandSerdeTests {
 
         byte[] serialised = serdes.commandId().serializer().serialize(topic, responseKey);
         CommandId deserialised = serdes.commandId().deserializer().deserialize(topic, serialised);
-        assertThat(deserialised).isEqualToComparingFieldByField(responseKey);
+        assertThat(deserialised).isEqualTo(responseKey);
+    }
+
+    @Test
+    void aggregateUpdate() {
+        AggregateUpdate<Boolean> update = new AggregateUpdate<>(true,
+                Sequence.first());
+
+        byte[] serialised = serdes.aggregateUpdate().serializer().serialize(topic, update);
+        AggregateUpdate<Boolean> deserialised = serdes.aggregateUpdate().deserializer().deserialize(topic, serialised);
+        assertThat(deserialised).isEqualToComparingFieldByField(update);
     }
 
     @Test
     void commandRequest() {
         UserAccountDomainKey aggKey = new UserAccountDomainKey("userId");
 
-        CommandRequest<UserAccountDomainKey, UserAccountDomainCommand> commandRequest = CommandRequest.of(
+        CommandRequest<UserAccountDomainKey, UserAccountDomainEvent> commandRequest = CommandRequest.of(
                 CommandId.random(),
                 aggKey,
                 Sequence.first(),
-                new UserAccountDomainCommand.UpdateUserName("name"));
+                new UserAccountDomainEvent.AccountCreated("name", Money.valueOf("100")));
 
         byte[] serialised = serdes.commandRequest().serializer().serialize(topic, commandRequest);
-        CommandRequest<UserAccountDomainKey, UserAccountDomainCommand> deserialised = serdes.commandRequest().deserializer().deserialize(topic, serialised);
+        CommandRequest<UserAccountDomainKey, UserAccountDomainEvent> deserialised = serdes.commandRequest().deserializer().deserialize(topic, serialised);
         assertThat(deserialised).isEqualToComparingFieldByField(commandRequest);
+    }
+
+    @Test
+    void eventWithSequence() {
+        ValueWithSequence<UserAccountDomainEvent> eventSeq = new ValueWithSequence<>(
+                new UserAccountDomainEvent.AccountCreated("name", Money.valueOf("100")),
+                Sequence.first()                );
+
+        byte[] serialised = serdes.valueWithSequence().serializer().serialize(topic, eventSeq);
+        ValueWithSequence<UserAccountDomainEvent> deserialised = serdes.valueWithSequence().deserializer().deserialize(topic, serialised);
+        assertThat(deserialised).isEqualToComparingFieldByField(eventSeq);
     }
 
     @Test
     void commandResponseSuccess() {
         UserAccountDomainKey aggKey = new UserAccountDomainKey("userId");
-
         CommandResponse commandResponse = CommandResponse.of(
                 CommandId.random(),
                 aggKey,
@@ -78,7 +101,6 @@ class AvroCommandSerdeTests {
     @Test
     void commandResponseFailure() {
         UserAccountDomainKey aggKey = new UserAccountDomainKey("userId");
-
         CommandResponse commandResponse = CommandResponse.of(
                 CommandId.random(),
                 aggKey,
@@ -89,4 +111,5 @@ class AvroCommandSerdeTests {
         CommandResponse deserialised = serdes.commandResponse().deserializer().deserialize(topic, serialised);
         assertThat(deserialised).isEqualToComparingFieldByField(commandResponse);
     }
+
 }
