+++ simplesource-command-serialization/src/test/resources/UserAccount.avdl
@@ -5,7 +5,7 @@ protocol UserSubsystem {
     string id;
   }
 
-  // 1. Define a basic aggregate model
+  // 1. Define a basic aggregateOf model
   record UserAccount {
     string name;
     decimal(12,4) balance;
